systemd: Flip -D userdb to true
GDM 49 will use it.

There seems no additional dependencies here so we need to edit LFS
anyway (at least we'll need to remove the incorrect "with dependencies
that do not fit in LFS").  And the increased SBU is negligiable, and
enabling it in LFS (instead of BLFS) has an advantage that "systemctl
preset-all" will automatically enable it.  Thus adding it in LFS
instead of only BLFS.