gcc: Fix the rationale not to enable more languages
The prior rationale ("missing dependencies") is just wrong and it is
confusing people in a heated discussion.  No matter how the discussion
will conclude, we should provide the correct information anyway.

Also document the --disable-bootstrap option that we'd not documented
elsewhere.

Now technically we may remove the BLFS GCC page as all its info is
already here, but doing so would be highly debatable.